requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.1
tzdata==2025.2
wheel==0.45.1
django-tailwind
tailwind
django-redis==5.4.0
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0